Fix Potion and Arrow PVP bypasses This is done by making a more defined check for making sure if the event is fired during a PVP event or not. This can be seen in the Utils.isPvP boolean more clearly. It essentially works by checking if both parties in the damage event are players, or if one of them is a projectile. If one is a projectile, it gets the shooter and checks if *that* is a player, then cancels the event properly.
For the potion splashes, it's pretty self-explanatory. They're fixed by listening to the PotionSplashEvent, determining if players are involved, and if so it loops through the potion effects, then loops through the affected entities, and if one of the entities is player with their pvp disabled, it sets the potion intensity to 0.