UnitedBrands - v3.1 Minor release here changing the way stat saving works. Previously, it used to save all the data in one yml file. While this approach works fine, its not very scalable. And considering the fact that any player who's ever made a brew will get registered in there (with all their stat nodes), it'd eventually get laggy when the server grows in the future.
This release makes the plugin more scalable by saving each player's data in their own file, and only creating the necessary nodes when needed. It's almost the exact same structure as UnitedPvP or UnitedChat.

This also fixes some bugs with the /brewery command not working properly after the code cleanup in the last version