UnitedSkills - v2.16 Point Management Now that all the skills are done and mostly balanced, we can move onto implementing the actual system with the jobs plugin. This commit brings in a point management system for the plugin.
When a player joins for the first time, all their points are initialized to 0. If they've already played before, the points are set according to what their current level in each job is. When a player levels up, one point is added to the respective job which they levelled up in.

There's a new /points command, which helps us manage points. The syntax is pretty simple, as its just /points add/remove <player> <job name> <amount>

This also brings in minor code cleanup in the registerListeners method in the main class.