UnitedItems - v2.0 Big release here! First of all, this commit axes out MOST features in the plugin in favor of itemsadder. We now explicitly handle the tree-side of things in our custom item system, and all the other parts of the system are to be moved to itemsadder to make this plugin lighter, and allow us much more flexibility with how we handle the actual items.
The plugin now has its own set of saplings, and when planted, those saplings are added into a little database. When a new tree grows, we check if the sapling in question is registered in our database, and if it is, we replace all the log and leaf blocks with their respective ItemsAdder block type.